Create FunctionMeta to store metadata for functions.
This allows us to stop holding state information on functions themselves
(aka stop treating functions as a "bag of properties").

Once this PR lands we will still be storing these function meta objects
on the function itself (in a `__ember_function_meta__` property) just
like we store normal object meta (in `__ember_meta__` property).  In the
longer term we will be able to avoid that also once we start storing meta
in a native WeakMap.